+++ prep_instruments.py
@@ -1,12 +1,15 @@
 import pyvisa
 import time
+import logging
+
+log = logging.getLogger(__name__)
 
 VISA_ADDRESS_POWER_METER    = 'USB0::0x0957::0x3718::DE53500131::0::INSTR'
 VISA_ADDRESS_TLS            = 'TCPIP0::100.65.2.45::inst0::INSTR'       
 
 def exceptionHandler(exception):
 
-    print('Error information:\n\tAbbreviation: %s\n\tError code: %s\n\tDescription: %s' % \
+    log.error('Error information:\n\tAbbreviation: %s\n\tError code: %s\n\tDescription: %s' % \
           (exception.abbreviation, exception.error_code, exception.description))
 
 def get_inst():
@@ -17,35 +20,35 @@ def get_inst():
         pm    = rm.open_resource(VISA_ADDRESS_POWER_METER)
         laser = rm.open_resource(VISA_ADDRESS_TLS)
     except visa.VisaIOError as ex:
-        print('VISA Error')
+        log.error('VISA Error')
         exceptionHandler(ex)
 
     pm.read_termination    = '\n'
     laser.read_termination = '\n'   
 
-    print(pm.query("*IDN?"))
-    print(laser.query("*IDN?"))
+    log.info(pm.query("*IDN?"))
+    log.info(laser.query("*IDN?"))
 
-    print("Connected\n")
+    log.info("Connected\n")
     return pm, laser
 
 
 def reset_inst(pm, laser):
-    print("[PM] Reset")
+    log.info("[PM] Reset")
     pm.write("*RST")
     
-    print("[LASER] Reset")
+    log.info("[LASER] Reset")
     laser.write("*RST")
     time.sleep(3)
 
 
 def init_inst(pm, laser):
-    print("Initializing instruments...")
+    log.info("Initializing instruments...")
     reset_inst(pm, laser)
 
     # Laser
     if (int(laser.query(":LOCK?")) == 1):
-        print("[LASER] TLS Locked. Unlock")
+        log.warning("[LASER] TLS Locked. Unlock")
         laser.write(":LOCK 0, 1234")
 
     laser.write(":STAT:QUES:ENAB 32767")    
@@ -53,7 +56,7 @@ def init_inst(pm, laser):
     # Power Meter
     pm.write(":STAT:QUES:ENAB 32767")
 
-    print("Instruments ready")
+    log.info("Instruments ready")
 
 
 def prep_inst():